Correct chunking issue that affected Yaml Doc parsing.
The chunking function was not aware of EOF, so when files did not end with a
separator, its last YamlDoc would be merged with the first YamlDoc of the next.

This resolves the issue by manually adding a separator between each file. This
is fine for case where there is already a separator, as empty YamlDocs ares
ignored.